+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.9.10] - UNRELEASED
+
+- Added ability to ignore patching meshes by mod in the mod dialog
+- Mods that contain just meshes will also be in the mod dialog now
+- UX improvements to Mod dialog
+
 ## [0.9.9] - 2025-12-26
 
 - Added warnings and error messages to confirmation dialog with ability to ignore them persistently